WIP: Refactor builder
M	include/vara/Feature/Feature.h
M	include/vara/Feature/FeatureModel.h
A	include/vara/Feature/FeatureModelBuilder.h
M	include/vara/Feature/FeatureModelParser.h
M	include/vara/Feature/FeatureModelTransaction.h
M	lib/Feature/CMakeLists.txt
M	lib/Feature/FeatureModel.cpp
A	lib/Feature/FeatureModelBuilder.cpp
M	lib/Feature/FeatureModelParser.cpp
M	unittests/Feature/BinaryFeature.cpp
M	unittests/Feature/FeatureModel.cpp
M	unittests/Feature/FeatureModelBuilder.cpp
M	unittests/Feature/FeatureModelTransaction.cpp
M	unittests/Feature/NumericFeature.cpp
M	unittests/Feature/OrderedFeatureVector.cpp
M	unittests/Feature/Relationship.cpp
M	unittests/Feature/RootFeature.cpp
