+++ include/vara/Feature/FeatureModel.h
@@ -39,7 +39,7 @@ public:
 
   FeatureModel(std::string Name, fs::path RootPath, std::string Commit,
                FeatureMapTy Features, ConstraintContainerTy Constraints,
-               RelationshipContainerTy Relationships, Feature *Root)
+               RelationshipContainerTy Relationships, RootFeature *Root)
       : Name(std::move(Name)), Path(std::move(RootPath)),
         Commit(std::move(Commit)), Features(std::move(Features)),
         Constraints(std::move(Constraints)),
@@ -116,7 +116,7 @@ protected:
   FeatureMapTy Features;
   ConstraintContainerTy Constraints;
   RelationshipContainerTy Relationships;
-  Feature *Root{nullptr};
+  RootFeature *Root{nullptr};
 
   FeatureModel() = default;
 
@@ -131,146 +131,12 @@ private:
   /// Delete a \a Feature.
   void removeFeature(Feature &Feature);
 
+  RootFeature *setRoot(std::unique_ptr<RootFeature> NewRoot);
+
   OrderedFeatureTy OrderedFeatures;
 };
 
-//===----------------------------------------------------------------------===//
-//                     Builder for FeatureModel
-//===----------------------------------------------------------------------===//
-
-/// \brief Builder for \a FeatureModel which can be used while parsing.
-class FeatureModelBuilder : private FeatureModel {
-public:
-  FeatureModelBuilder() = default;
-  void init() {
-    Name = "";
-    Path = "";
-    Commit = "";
-    Root = nullptr;
-    Features.clear();
-    Constraints.clear();
-    Parents.clear();
-    Children.clear();
-  }
-
-  /// Try to create a new \a Feature.
-  ///
-  /// \param[in] FeatureName name of the \a Feature
-  /// \param[in] FurtherArgs further arguments that should be passed to the
-  ///                        \a Feature constructor
-  ///
-  /// \returns ptr to inserted \a Feature
-  template <typename FeatureTy, typename... Args,
-            std::enable_if_t<std::is_base_of_v<Feature, FeatureTy>, int> = 0>
-  FeatureTy *makeFeature(std::string FeatureName, Args &&...FurtherArgs) {
-    if (!Features
-             .try_emplace(FeatureName,
-                          std::make_unique<FeatureTy>(
-                              FeatureName, std::forward<Args>(FurtherArgs)...))
-             .second) {
-      return nullptr;
-    }
-    return llvm::dyn_cast<FeatureTy>(Features[FeatureName].get());
-  }
-
-  FeatureModelBuilder *addEdge(const std::string &ParentName,
-                               const std::string &FeatureName) {
-    Children[ParentName].insert(FeatureName);
-    Parents[FeatureName] = ParentName;
-    return this;
-  }
-
-  FeatureModelBuilder *
-  emplaceRelationship(Relationship::RelationshipKind RK,
-                      const std::vector<std::string> &FeatureNames,
-                      const std::string &ParentName) {
-    RelationshipEdges[ParentName].emplace_back(RK, FeatureNames);
-    return this;
-  }
-
-  FeatureModelBuilder *
-  addConstraint(std::unique_ptr<FeatureModel::ConstraintTy> C) {
-    Constraints.push_back(std::move(C));
-    return this;
-  }
-
-  FeatureModelBuilder *setVmName(std::string Name) {
-    this->Name = std::move(Name);
-    return this;
-  }
-
-  FeatureModelBuilder *setPath(fs::path Path) {
-    this->Path = std::move(Path);
-    return this;
-  }
-
-  FeatureModelBuilder *setCommit(std::string Commit) {
-    this->Commit = std::move(Commit);
-    return this;
-  }
-
-  FeatureModelBuilder *setRootName(std::string Name) {
-    this->RootName = std::move(Name);
-    return this;
-  }
-
-  /// Build \a FeatureModel.
-  ///
-  /// \return instance of \a FeatureModel
-  std::unique_ptr<FeatureModel> buildFeatureModel();
-
-private:
-  class BuilderVisitor : public ConstraintVisitor {
-
-  public:
-    BuilderVisitor(FeatureModelBuilder *Builder) : Builder(Builder) {}
-
-    void visit(PrimaryFeatureConstraint *C) override {
-      auto *F = Builder->getFeature(C->getFeature()->getName());
-      C->setFeature(F);
-      F->addConstraint(C);
-    };
-
-  private:
-    FeatureModelBuilder *Builder;
-  };
-
-  using EdgeMapType = typename llvm::StringMap<llvm::SmallSet<std::string, 3>>;
-  using RelationshipEdgeType = typename llvm::StringMap<std::vector<
-      std::pair<Relationship::RelationshipKind, std::vector<std::string>>>>;
-
-  FeatureModel::ConstraintContainerTy Constraints;
-  FeatureModel::RelationshipContainerTy Relationships;
-  llvm::StringMap<std::string> Parents;
-  EdgeMapType Children;
-  RelationshipEdgeType RelationshipEdges;
-  std::string RootName{"root"};
-
-  bool buildRoot();
-
-  bool buildConstraints();
-
-  /// This method is solely relevant for parsing XML, as alternatives are
-  /// represented als mutual excluded but non-optional features (which requires
-  /// additional processing).
-  void detectXMLAlternatives();
-
-  bool buildTree(const std::string &FeatureName,
-                 std::set<std::string> &Visited);
-};
 } // namespace vara::feature
-
-inline std::ostream &operator<<(std::ostream &Out,
-                                const vara::feature::Feature *Feature) {
-  Out << Feature->toString();
-  return Out;
-}
-
-inline llvm::raw_ostream &operator<<(llvm::raw_ostream &Out,
-                                     const vara::feature::Feature *Feature) {
-  Out << Feature->toString();
-  return Out;
-}
 
 namespace llvm {
