+++ unittests/Feature/OrderedFeatureVector.cpp
@@ -1,4 +1,4 @@
-#include "vara/Feature/FeatureModel.h"
+#include "vara/Feature/FeatureModelBuilder.h"
 
 #include "gtest/gtest.h"
